Fix useradd call and ensure lazy map eval happens
In python 3 map is lazy evaluated, so if you don't iterate it, then you
will never execute the function. So this casting it to a list will step
through the iterator.

The useradd call didn't have the needed params, first it needed to not
set any of the user data since that is asked in stdin. It could
potentially be grabbed from AWS IAM.

The shell needs to be set so they are actually dropped into a shell when
they try to login.

Disabling the password prompt since they will login with keys.